requirements.txt
whisper
moviepy
kivy
Pillow
ffmpeg-python
transformers